requirements/app.txt
aiohttp==3.5.4
galaxy.plugin.api==0.47
certifi==2019.3.9
psutil==5.6.1
pyobjc-core==5.2; sys_platform == 'darwin'
pyobjc-framework-Cocoa==5.2; sys_platform == 'darwin'